Add physics step counter to `PhysicsTotalDiagnostics`
